fix(gates): split sloppy class-field stores from property_set.rs (file cap)
